Split test cases into two files
