+++ .vscode/extensions.json
@@ -1,13 +1,11 @@
 {
-    "recommendations": [
-        "bierner.github-markdown-preview",
-        "dbaeumer.vscode-eslint",
-        "eg2.vscode-npm-script",
-        "esbenp.prettier-vscode",
-        "hbenl.vscode-mocha-test-adapter",
-        "hbenl.vscode-test-explorer",
-        "redhat.vscode-xml",
-        "huntertran.auto-markdown-toc",
-        "streetsidesoftware.code-spell-checker"
-    ]
+  "recommendations": [
+    "bierner.github-markdown-preview",
+    "dbaeumer.vscode-eslint",
+    "esbenp.prettier-vscode",
+    "hbenl.vscode-mocha-test-adapter",
+    "hbenl.vscode-test-explorer",
+    "redhat.vscode-xml",
+    "streetsidesoftware.code-spell-checker"
+  ]
 }
